requirements.txt
pyTelegramBotAPI==4.5.1
python-dotenv==0.20.0
requests==2.27.1
